requirements.txt
tk
opencv-python
numpy
mediapipe
autopy
pyautogui
wmi
Pillow
pycaw
comtypes
pyinstaller
